feat(config): split SiliconFlow into CN and global variants
Add support for api.siliconflow.com domain alongside the existing
api.siliconflow.cn. Replace SVG logo with PNG format.
